Add CleanPlayer component tests
